requirements.txt
crosshair-tool
icontract
argparse
mpmath
sympy
regex
django
coverage